+++ ui/src/kapacitor/containers/KapacitorRulePage.js
@@ -18,8 +18,6 @@ class KapacitorRulePage extends Component {
       enabledAlerts: [],
       kapacitor: {},
     }
-
-    this.isEditing = ::this.isEditing
   }
 
   async componentDidMount() {
@@ -97,7 +95,7 @@ class KapacitorRulePage extends Component {
     )
   }
 
-  isEditing() {
+  isEditing = () => {
     const {params} = this.props
     return params.ruleID && params.ruleID !== 'new'
   }